Add Sklearn Training Operator  (#2676)
## Required packages

- sklearn

## KNN

This operator is used to train the dataset in KNN model. There are two
modes for this operator, which are KNN classifier and regressor, and
they have same input ports, output ports, and parameter. This operator
aims to train the model, and pass the model information to the next
operator.

### Properties
<img width="253" alt="image"
src="https://github.com/Texera/texera/assets/60329261/34c9ddb0-7840-4b7b-813c-e3926046433d">

### Input port

**Dataset**

Training dataset

**Parameter**

Parameter of model

### Output port

Trained Model

### Demo

![20240526223804](https://github.com/Texera/texera/assets/60329261/f3121a8a-97c8-4666-9e1a-661819721da2)

![20240526223925](https://github.com/Texera/texera/assets/60329261/4088843d-4051-4b10-a096-64f221a98def)

---------

Co-authored-by: Jiayi Huang <huangjiayi@JiayideAir.lan>
Co-authored-by: Harry Xiong <harryxiong24@gmail.com>
Co-authored-by: Jiayi Huang <huangjiayi@dhcp-172-31-142-129.mobile.uci.edu>
Co-authored-by: Jiayi Huang <huangjiayi@JiayideMacBook-Air.local>
Co-authored-by: Xinyuan Lin <xinyual3@uci.edu>