Fix jump back button does not work properly on workflow detail page (#3026)
### Purpose:
Sometimes, the Jump Back button on the Hub Workflow Detail page does not
work properly. This PR fixes the issue. The root cause of the problem is
that when a user clicks on an operator in the preview, the URL of the
Detail page would slightly changes. In the previous version, the Jump
Back button relied on the `back()` method, which navigates based on
browser history.
fix #3015

### Changes:
1. Directly use `navigateByUrl()` to return to the Hub Workflow page.
2. Create an `app-routing.constants.ts` file to store all the addresses
used in the frontend as constants.

### Demo:
URL after click operator:

![image](https://github.com/user-attachments/assets/feb21c18-9b73-4612-a8ec-b9412392bee2)

Before Changes:


https://github.com/user-attachments/assets/4682ac46-0a5e-4a02-89f1-04d15cecd400


After Changes:


https://github.com/user-attachments/assets/d2dc2a31-daf5-4ee5-aca8-6822bd94dc77