Fix Sink Operator Storage Flush Logic (#3060)
This PR fixes #3056. The reason for the random behavior of #3056 is: 
- There is an additional materialization pairs added after `DISTINCT`,
and the next operator would belong to a new region.
- The current sink operator implementation, which does batch write, only
flushes the storage writer when the operator closes.
- The completion of the region that a sink operator belongs to is marked
by the completion its input port. This causes the next region to be
started as soon as its the input port is finished, and there is a chance
that the materialized write is not flushed yet.
- The fix is to move the storage flush to as soon as the input port of
the sink is finished.