Allow force casting of comma-separated numbers (#3059)
The typecasting operator does not work when casting number-strings with
commas. This PR adds this functionality. By default, it will infer the
comma-separated number-strings as String type to preserve the
information. When the user explicitly requests casting such a
comma-separated number-string into an Int type or a Long type, it will
perform the casting.

For instance, now typecasting can cast `1,234` into `1234` as an
integer.

Added testcases for other types as well.