Fix result panel docking behavior (#3030)
## Purpose
Address #2921 
The docking logic for the panels does not consider the result panel's
resizing. As a result, the result panel could be resized from the bottom
and bottom right when docked, so if the panel gets smaller or bigger,
there is either leftover space or the panel goes over the screen. This
PR aims to address this issue
## Changes
Disable result-panel resize from the bottom and bottom-right when the
panel is docked.
Make changes in result-panel's docking logic to account for resizes, so
that there is no leftover space or the panel doesn't go over the screen
when pressing docking button.
## Demo
Before

![chrome_DKjeyyQxQD](https://github.com/user-attachments/assets/f13f2dfe-023c-4d13-b152-b3cd4cb52059)
After

https://github.com/user-attachments/assets/95658075-a4da-472e-81d1-eccfe7da639a